Added support for zipped bags (#21)
* NO JIRA. Added support for zipped bags
* Fixed some docs and var names
* Corrected docs for run.sh

+++ src/main/java/nl/knaw/dans/easy/sword2examples/ContinuedDeposit.java
@@ -30,7 +30,7 @@ import java.security.MessageDigest;
 public class ContinuedDeposit {
     public static void main(String[] args) throws Exception {
         if (args.length != 5) {
-            System.err.println("Usage: java nl.knaw.dans.easy.sword2examples.ContinuedDeposit <Col-IRI> <EASY uid> <EASY passwd> <chunk size> <bag dirname>");
+            System.err.println("Usage: java nl.knaw.dans.easy.sword2examples.ContinuedDeposit <Col-IRI> <EASY uid> <EASY passwd> <chunk size> <bag file/dir>");
             System.exit(1);
         }
 
@@ -39,16 +39,20 @@ public class ContinuedDeposit {
         final String uid = args[1];
         final String pw = args[2];
         final int chunkSize = Integer.parseInt(args[3]);
-        final String bagFileName = args[4];
+        final String bagFile = args[4];
 
-        depositPackage(new File(bagFileName), colIri, uid, pw, chunkSize);
+        depositPackage(new File(bagFile), colIri, uid, pw, chunkSize);
     }
 
-    public static URI depositPackage(File bagDir, IRI colIri, String uid, String pw, int chunkSize) throws Exception {
-        // 0. Zip the bagDir
-        File zipFile = new File(bagDir.getAbsolutePath() + ".zip");
-        zipFile.delete();
-        Common.zipDirectory(bagDir, zipFile);
+    public static URI depositPackage(File bag, IRI colIri, String uid, String pw, int chunkSize) throws Exception {
+        File zipFile = null;
+        if (bag.isDirectory()) {
+            zipFile = new File(bag.getAbsolutePath() + ".zip");
+            zipFile.delete();
+            Common.zipDirectory(bag, zipFile);
+        } else {
+            zipFile = bag;
+        }
 
         // 1. Set up stream for calculating MD5
         FileInputStream fis = new FileInputStream(zipFile);
@@ -75,7 +79,7 @@ public class ContinuedDeposit {
         Link seIriLink = receipt.getLink("edit");
         URI seIri = seIriLink.getHref().toURI();
 
-        int remaining = (int) zipFile.length() - chunkSize;
+        long remaining = zipFile.length() - chunkSize;
         int count = 2;
         while (remaining > 0) {
             System.out.print(String.format("POST-ing chunk of %d bytes to SE-IRI (remaining: %d) ... ", chunkSize, remaining));
